crud_pets/requirements.txt
appnope==0.1.0
backcall==0.1.0
blinker==1.4
certifi==2018.4.16
click==6.7
decorator==4.3.0
flake8==3.5.0
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-Modus==0.0.1
Flask-SQLAlchemy==2.3.2
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
parso==0.2.1
pbr==4.0.4
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.8
stevedore==1.28.0
traitlets==4.3.2
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
Werkzeug==0.14.1
yapf==0.22.0
